✨ feat(user): 기수별 사용자 조회 및 이메일 기반 삭제 추가

+++ src/main/java/com/pard/pard_backend/domain/ranking/dto/response/RankingResponseDTO.java
@@ -17,7 +17,7 @@ public class RankingResponseDTO {
         return RankingResponseDTO.builder()
                 .name(user.getName())
                 .part(user.getPart())
-                .totalBonus(user.getTotalBonus())
+                .totalBonus((int)user.getTotalBonus())
                 .build();
     }
